Game Complete
I finished Micro Game 2.
Implemented a scrolling background.
Added in a Player space ship with a controller which controls vertical position on the screen and fires lasers with the space bar.
Added an Enemy spawner which creates an enemy ship every two seconds which shoots lasers.
Added collisions so when a laser hits another laser or an enemy it is destroyed.
The life system has hearts that disappear every time you are hit.