Micro Game #4
Added features onto Micro Game 2.
Pause button added into the main scene.
Added feature that speeds up the spawn rate of enemies the more points you get. Max spawn rate is 0.5f.
Added Horizontal and Vertical Movement with Boundaries to the ship.
Added asteroids to the enemy spawner. The asteroids travel at a faster speed.
Added in game Score counter and improved the Health system.
Added a Main menu Screen with a play button.
Added a game over screen that displays the score and has a button to return to the main menu.